backend/requirements.txt
httpx>=0.27,<1
anthropic>=0.40,<1
pyjwt>=2.8,<3
pytest>=8,<9
pytest-asyncio>=0.24,<1
respx>=0.22,<1
